+++ src/instances/usd-formatter.ts
@@ -1,9 +1,0 @@
-//
-// Instance
-//
-
-export const usdFormatter = new Intl.NumberFormat("en-US",
-	{
-		style: "currency",
-		currency: "USD",
-	});